api_yamdb/requirements.txt
asgiref==3.2.10
Django==2.2.16
djangorestframework==3.12.4
django-filter==2.2.0
djangorestframework-simplejwt==4.7.2
PyJWT==2.1.0
requests==2.26.0
pytz==2020.1
sqlparse==0.3.1 
gunicorn==20.0.4
#psycopg2-binary==2.8.6
